Updated 'MemCopy' to call RtlMoveMemory directly for sizes > 16777216 bytes
Speed benchmarks clearly show that after a certain size the overhead of the API call (issue in VBA7 only) is negligible and the API call is faster than using BSTR/SAFEARRAY to copy memory. This also fixes an issue with overlapping memory when copying in chunks with SAFEARRAY - no more chunks now.