Fixed bug in 'CloneParamArray' and updated signature
If the first element of the param array was a Variant that was ByRef then the whole memory copy would have happened at the address of the Variant instead of the address of the first element in the param array. The only way to get the correct address is to retrieve the address of the actual SAFEARRAY and because a param array cannot be coerced into a variant then the only way to get the address is to use the 'Not Not paramArr' trick.